backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.10.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.0.0
websockets>=11.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.0
